# Improve performance of `WP_Theme_JSON::get_merged_data` method