@oandregal 2022-10-13
cc @ockham @hellofromtonya this is ready for review/merge.

There are some failures in the `PHPUnit Tests / 8.2 on ubuntu-latest` job. These are totally unrelated to this PR and I don't know why they fail. They say things such as:

```bash
For automated test runs, this test is only run on trunk
This test requires the index to be truncated.
Rendering PDFs is not supported on this system.
The image editor engine WP_Image_Editor_Imagick is not supported on this system.
```

I'd appreciate if someone could re-run these or provide help on what's happening to unblock this PR.