@oandregal 2022-10-14
Rebased from `trunk` to address the [environment failures](https://github.com/WordPress/wordpress-develop/pull/3441#issuecomment-1277450187). I've seen some test have been disabled since this was reported.