@jorgefilipecosta 2022-10-14
Committed to trunk in https://github.com/WordPress/wordpress-develop/commit/cb98544ed1fe76e62823a9397087433215f8f3f2. We should leave the PR open until there is a 6.1 branch and we confirm the commit is there (because the branch was created after the commit) or the commit is not there, but we do the backport.