@oandregal 2023-02-28
I'm proposing we revert this change at https://github.com/WordPress/wordpress-develop/pull/4145 as it introduced a bug.